관리자의 일정 승인 또는 반려

+++ src/main/java/com/atomz/sawonz/domain/user/dto/AdminDto.java
@@ -37,8 +37,8 @@ public class AdminDto {
         @Email
         private String email;
 
-        private Integer salary;
-        private Integer annualLeaveCount;
+        private Long salary;
+        private Double annualLeaveCount;
         private String positionTitle;
         private LocalDate hiredAt;
